requirements.txt
django
django-environ
djangorestframework
djangorestframework-simplejwt
djoser
psycopg2
